Do version upgrade.

Add sf_internal.h file to zynq_qspi.
Add zynqmp line to arch/arm/cpu/armv8/Makefile.

+++ arch/blackfin/cpu/jtag-console.c
@@ -168,7 +168,7 @@ static int jtag_getc(struct stdio_dev *dev)
 		inbound_len = emudat;
 	} else {
 		/* store the bytes */
-		leftovers_len = min(4, inbound_len);
+		leftovers_len = min((size_t)4, inbound_len);
 		inbound_len -= leftovers_len;
 		leftovers = emudat;
 	}
